add design

+++ src/app/profile/profile.component.ts
@@ -15,7 +15,7 @@ export class ProfileComponent implements OnInit {
   profilePicture:string='https://t4.ftcdn.net/jpg/05/89/93/27/360_F_589932782_vQAEAZhHnq1QCGu5ikwrYaQD0Mmurm0N.webp'
   userData:any={}
   suggestItem:any[]=[]
-  // lastModifyImg:any=''
+  
   constructor(private api:ApiCallService,private toastr:ToastrService,private route:Router){}
 
   ngOnInit() {
@@ -27,7 +27,7 @@ export class ProfileComponent implements OnInit {
   getUser(){
     this.api.getUserProfile().subscribe((res:any)=>{
       this.userData=res
-      // console.log("profile=",this.userData);
+      console.log("profile=",this.userData);
       if(this.userData.profileImage){
         this.profilePicture=res.profileImage
       }
@@ -61,31 +61,6 @@ export class ProfileComponent implements OnInit {
     }
   }
 
-  // handleUpdateUser(){     
-  //   this.api.updateUserProfile(this.userData).subscribe((res:any)=>{
-  //     this.toastr.success("Profile Updated Successfully!!")
-  //   },
-  //   (err:any)=>{
-  //     this.toastr.error("Updation Faild",err)
-  //     console.log(err);
-      
-  //   }
-  // )
-  // }
-
-  // handleUpdateAdmin(){     
-  //   this.api.updateAdminProfile(this.userData).subscribe((res:any)=>{
-  //     this.toastr.success("Profile Updated Successfully!!")
-  //   },
-  //   (err:any)=>{
-  //     this.toastr.error("Updation Faild",err)
-  //     console.log(err);
-      
-  //   }
-  // )
-  // }
-
-
   handleUpdateProfile() {
     if (sessionStorage.getItem("existingUser")) {
       this.api.updateUserProfile(this.userData).subscribe(
@@ -115,28 +90,36 @@ export class ProfileComponent implements OnInit {
   
   
   deleteAccount(){
-    this.api.deteteAccountApi().subscribe({
-      next:(res:any)=>{
-        console.log(res);
-        this.api.getCartListCountApi()
-        this.api.getWishlistCountApi()
-        this.getUser()
-        sessionStorage.clear()
-        localStorage.clear()
-        this.toastr.success("Account deleted SuccessFully!!")
-        this.route.navigateByUrl('/')
-      },
-      error:(err)=>{
-        console.log(err);
-        this.toastr.error("Account Deletion Faild!!")
-      }
-    })
+    const existingAdmin =sessionStorage.getItem("existingUser");
+    const role= sessionStorage.getItem("role")
+    if(existingAdmin && role ==="user"){
+      this.api.deteteAccountApi().subscribe({
+        next:(res:any)=>{
+          console.log(res);
+          this.api.getCartListCountApi()
+          this.api.getWishlistCountApi()
+          this.getUser()
+          sessionStorage.clear()
+          localStorage.clear()
+          this.toastr.success("Account deleted SuccessFully!!")
+          this.route.navigateByUrl('/')
+        },
+        error:(err)=>{
+          console.log(err);
+          this.toastr.error("Account Deletion Faild!!")
+        }
+      })
+    }
+    else{
+      this.toastr.info("Your an admin")
+    }
+    
   }
 
   getRecent(){
     this.api.getrecentProducts().subscribe((res:any)=>{
       this.suggestItem=res
-      console.log("Recent",res);
+      // console.log("Recent",res);
       
     })
   }
@@ -144,8 +127,7 @@ export class ProfileComponent implements OnInit {
   logout() { 
     sessionStorage.clear();
     localStorage.clear();
-    // this.wishCount = 0;
-    // this.cartCount = 0;
+    this.toastr.success("Logout successfully!!")
     this.getUser()
     this.route.navigateByUrl('/log')
 }
